requirements.txt
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
click==8.1.3
decorator==5.1.1
executing==1.2.0
Flask==2.3.1
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
libvirt-python==9.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
wcwidth==0.2.6
Werkzeug==2.3.2
